moved question redirect list to Question list
