+++ src/app/domains/Condition/components/ConditionForm/ConditionTemplates/YesNoChoiceTemplate/YesNoChoiceTemplate.js
@@ -2,15 +2,16 @@ import { Box, Col, Row } from '@qonsoll/react-design'
 import { Button, Select } from 'antd'
 import Text from 'antd/lib/typography/Text'
 import React, { useState } from 'react'
+import { QuestionSelect } from 'domains/Question/components'
 
 const { Option, OptGroup } = Select
 
 function YesNoChoiceTemplate(props) {
   const { answers, questionList, addRedirectQuestion } = props
 
-  const onChange = (question, index) => {
-    addRedirectQuestion(question, index)
-  }
+  // const onChange = (question, index) => {
+  //   addRedirectQuestion(question, index)
+  // }
   return (
     <>
       {answers.map((item, index) => (
@@ -30,35 +31,12 @@ function YesNoChoiceTemplate(props) {
             </Box>
           </Col>
           <Col>
-            <Box display="flex" border="1px solid #bbbbbb" borderRadius="4px">
-              <Select
-                value={
-                  answers[index].redirectQuestion || 'Go to the next question'
-                }
-                showSearch
-                allowClear
-                bordered={false}
-                onChange={(name) => onChange(name, index)}
-                defaultValue="Go to the next question"
-                style={{
-                  width: '100%',
-                  height: '48px',
-                  display: 'flex',
-                  alignItems: 'center'
-                }}>
-                <Button>
-                  <Text strong>Submit form</Text>
-                </Button>
-
-                <OptGroup label="JUMP TO...">
-                  {questionList.map((item, index) => (
-                    <Option key={index} value={item.name} onClick={() => {}}>
-                      {item.name}
-                    </Option>
-                  ))}
-                </OptGroup>
-              </Select>
-            </Box>
+            <QuestionSelect
+              addRedirectQuestion={addRedirectQuestion}
+              answers={answers}
+              index={index}
+              questionList={questionList}
+            />
           </Col>
         </Row>
       ))}
