+++ src/app/domains/Question/components/QuestionSelect/index.js
@@ -1,0 +1,3 @@
+import QuestionSelect from './QuestionSelect'
+
+export default QuestionSelect
