+++ src/app/domains/Question/components/index.js
@@ -1,2 +1,3 @@
 export { default as QuestionAdvancedView } from './QuestionAdvancedView'
 export { default as QuestionForm } from './QuestionForm'
+export { default as QuestionSelect } from './QuestionSelect'
