+++ src/app/domains/Question/components/index.json
@@ -1,1 +1,1 @@
-["QuestionAdvancedView","QuestionSimpleView","QuestionsList","QuestionForm"]
+["QuestionAdvancedView","QuestionForm","QuestionSimpleView","QuestionsList","QuestionSelect"]
